[REFACTOR] Create Coupon Validation Class
